adding driver support for RCC and GPIO

+++ configure.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
 # Read in device family
-echo "Device families: " && ls proc
+echo "Device families: " && echo "" && ls proc && echo ""
+
 read -t 10 -p "Which device family? " DEV_FAM
 
 # Read in processor
-echo "Processors: " && ls proc/$DEV_FAM 2>/dev/null
+echo "Processors: " && ls proc/$DEV_FAM 2>/dev/null | sed 's/drivers//'
+echo ""
 if [ $? -eq 2 ]; then
 	echo "'$DEV_FAM' not found."
 	exit 1
@@ -20,7 +22,16 @@ if [ $? -eq 2 ]; then
 	exit 1
 fi
 
+# Verify drivers dir exists
+ls proc/$DEV_FAM/drivers >/dev/null 2>/dev/null
+if [ $? -eq 2 ]; then
+	echo "'$DEV_FAM/drivers' not found."
+	exit 1
+fi
+
 rm -f cfg
+rm -f src/drivers
 ln -s proc/$DEV_FAM/$PROC_CHOICE cfg
+ln -s ../proc/$DEV_FAM/drivers src/drivers
 make clean
 echo "Project configured for $DEV_FAM$PROC_CHOICE."
